combine commands and specify seconds
